% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{routes}
\alias{routes}
\title{Get Routes}
\usage{
routes(
  x,
  overview = "full",
  nc = 1,
  nq = 100,
  server,
  profile = "auto",
  costing_options = list()
)
}
\arguments{
\item{x}{dataframe with x_src, y_src, x_dst, y_dst in WGS84 coordinates
(EPSG:4326)}

\item{overview}{overview (FALSE, for no geometry or "full" for detailed geometry)}

\item{nc}{number of CPU cores}

\item{nq}{number of queries per chunk}

\item{server}{Valhalla server address}

\item{profile}{Valhalla profile (costing model)}

\item{costing_options}{list of options to use with the costing model
(see \url{https://valhalla.github.io/valhalla/api/turn-by-turn/api-reference/#costing-options}
or valh documentation)}
}
\value{
An sf object (if overview == "full") or a data.frame
(if overview == FALSE) is returned.
}
\description{
This function enables the computation of routes.
}
\examples{
\dontrun{
apt <- read.csv(system.file("csv/apotheke.csv", package = "valh"))
x <- cbind(apt[c(1:100,1:100,1:100, 1:100), c(2:3)],
           apt[c(rep(1,100), rep(10,100), rep(20,100), rep(30, 100)), c(2:3)])
server = "http://xxxxxxx:8002/"
r <- routes(x, nc = 8, nq = 50, server = server)
library(mapsf)
op <- par(mfrow = c(2,2))
mf_map(r[1:100,], col = "red")
mf_map(r[101:200,], col = "blue")
mf_map(r[201:300,], col = "darkgreen")
mf_map(r[301:400,], col = "purple")
par(op)
}
}
